+++ src/it/test_ifs.rs
@@ -41,6 +41,8 @@ pub mod test_toplevel_drag;
 pub mod test_toplevel_drag_manager;
 pub mod test_viewport;
 pub mod test_viewporter;
+pub mod test_virtual_keyboard;
+pub mod test_virtual_keyboard_manager;
 pub mod test_xdg_activation;
 pub mod test_xdg_activation_token;
 pub mod test_xdg_base;