+++ src/it/test_ifs/test_registry.rs
@@ -15,8 +15,9 @@ use {
                 test_single_pixel_buffer_manager::TestSinglePixelBufferManager,
                 test_subcompositor::TestSubcompositor, test_syncobj_manager::TestSyncobjManager,
                 test_toplevel_drag_manager::TestToplevelDragManager,
-                test_viewporter::TestViewporter, test_xdg_activation::TestXdgActivation,
-                test_xdg_base::TestXdgWmBase,
+                test_viewporter::TestViewporter,
+                test_virtual_keyboard_manager::TestVirtualKeyboardManager,
+                test_xdg_activation::TestXdgActivation, test_xdg_base::TestXdgWmBase,
             },
             test_object::TestObject,
             test_transport::TestTransport,
@@ -52,6 +53,7 @@ pub struct TestRegistrySingletons {
     pub zwp_linux_dmabuf_v1: u32,
     pub xdg_toplevel_drag_manager_v1: u32,
     pub wp_alpha_modifier_v1: u32,
+    pub zwp_virtual_keyboard_manager_v1: u32,
 }
 
 pub struct TestRegistry {
@@ -76,6 +78,7 @@ pub struct TestRegistry {
     pub dmabuf: CloneCell<Option<Rc<TestDmabuf>>>,
     pub drag_manager: CloneCell<Option<Rc<TestToplevelDragManager>>>,
     pub alpha_modifier: CloneCell<Option<Rc<TestAlphaModifier>>>,
+    pub virtual_keyboard_manager: CloneCell<Option<Rc<TestVirtualKeyboardManager>>>,
     pub seats: CopyHashMap<GlobalName, Rc<WlSeatGlobal>>,
 }
 
@@ -144,6 +147,7 @@ impl TestRegistry {
             zwp_linux_dmabuf_v1,
             xdg_toplevel_drag_manager_v1,
             wp_alpha_modifier_v1,
+            zwp_virtual_keyboard_manager_v1,
         };
         self.singletons.set(Some(singletons.clone()));
         Ok(singletons)
@@ -237,6 +241,13 @@ impl TestRegistry {
         wp_alpha_modifier_v1,
         1,
         TestAlphaModifier
+    );
+    create_singleton!(
+        get_virtual_keyboard_manager,
+        virtual_keyboard_manager,
+        zwp_virtual_keyboard_manager_v1,
+        1,
+        TestVirtualKeyboardManager
     );
 
     pub fn bind<O: TestObject>(