+++ src/it/test_transport.rs
@@ -69,6 +69,7 @@ impl TestTransport {
             dmabuf: Default::default(),
             drag_manager: Default::default(),
             alpha_modifier: Default::default(),
+            virtual_keyboard_manager: Default::default(),
             seats: Default::default(),
         });
         self.send(wl_display::GetRegistry {