+++ release-notes.md
@@ -2,6 +2,7 @@
 
 - Add support for wp-alpha-modifier.
 - Add support for per-device keymaps.
+- Add support for virtual-keyboard-unstable-v1.
 
 # 1.0.3 (2024-04-11)